Timeline to replace Time
Added a Timeline struct to manage constants and calculations related to the time represented onscreen.
Builds and runs so I'm saving it! Need to complete switching functionality over from the previous struct, Time. Partly this is to get the dependencies right; some UI elements should update when now changes, some when the trailingTime changes, and some when either changes.