Fixed Issue with Backdrop
Colors no longer wildly fluctuating at the right edge of the screen. I simplified the process for producing stops, factored out the interpolation function, and tightened up the logic so it's now very clear where and how the edge cases are handled. And it seems to work great. Issues with event views are not resolved - still tinkering with the structure of the app to make that stuff work well.