Expand Events Using Set of IDs
I re-factored the way events were expanded. We had a binding AND an array of indexes; changed to a set of IDs with no binding. Events are added to or removed from the set via methods on EventManager. Now it's consistent. However, sometimes closing an event closes one, sometimes all...